requirements.txt
py-solc-x
python-dotenv
web3
requests
